+++ src/main/java/com/movisens/smartgattlib/helper/GattByteBuffer.java
@@ -2,7 +2,6 @@ package com.movisens.smartgattlib.helper;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
-import java.time.ZoneId;
 import java.util.Date;
 import java.util.UUID;